requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.2
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.1
Pillow==8.0.1
pycodestyle==2.6.0
pycryptodome==3.9.9
pylint==2.6.0
pymongo==3.11.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2020.4
regex==2020.11.11
requests==2.25.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.1
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
